refactor(pipeline): centralize config and extract downloadWithRetry
- Move MLIT shape mapping (operator, lineToRouteId) from
  build-train-shapes.ts to toei-train resource definition
- Extract downloadWithRetry with progress display from download-gtfs.ts
  into shared download-utils.ts for reuse across download scripts
