+++ pipeline/lib/download-utils.ts
@@ -6,8 +6,10 @@
  */
 
 import { execFileSync } from 'node:child_process';
-import { existsSync, mkdirSync } from 'node:fs';
+import { createWriteStream, existsSync, mkdirSync } from 'node:fs';
 import { resolve } from 'node:path';
+import { Readable, Transform } from 'node:stream';
+import { pipeline } from 'node:stream/promises';
 import { pathToFileURL } from 'node:url';
 
 // ---------------------------------------------------------------------------
@@ -144,6 +146,86 @@ export async function withRetry<T>(
   }
   // TypeScript control flow: loop always returns or throws, but compiler cannot prove it.
   throw new Error(`Failed after ${maxRetries} attempts (${label})`, { cause: lastError });
+}
+
+// ---------------------------------------------------------------------------
+// Download with retry and progress
+// ---------------------------------------------------------------------------
+
+/** Result of a successful file download. */
+export interface DownloadResult {
+  /** Total bytes written to disk. */
+  bytes: number;
+  /** Download duration in milliseconds. */
+  durationMs: number;
+  /** HTTP Content-Type header value, if present. */
+  contentType: string;
+}
+
+/**
+ * Download a file with retry, progress display, and streaming to disk.
+ *
+ * Shows real-time download progress (percentage when Content-Length is
+ * available, otherwise raw byte count). Retries on failure with
+ * exponential backoff via {@link withRetry}.
+ *
+ * @param url - URL to download.
+ * @param dest - Local file path to write to.
+ * @returns Download result with byte count, duration, and content type.
+ */
+export async function downloadWithRetry(url: string, dest: string): Promise<DownloadResult> {
+  return withRetry(async () => {
+    let res: Response;
+    const startTime = performance.now();
+    try {
+      res = await fetch(url, {
+        signal: AbortSignal.timeout(FETCH_TIMEOUT_MS),
+      });
+    } catch (err) {
+      throw wrapTimeoutError(err, url);
+    }
+    if (!res.ok) {
+      throw new Error(`HTTP ${res.status} ${res.statusText}`);
+    }
+
+    const contentLength = Number(res.headers.get('content-length') || 0);
+    const contentType = res.headers.get('content-type') ?? '';
+    if (contentLength > 0) {
+      console.log(`  Content-Length: ${contentLength.toLocaleString()} bytes`);
+    }
+
+    if (!res.body) {
+      throw new Error(`Response body is empty for ${url}`);
+    }
+
+    const out = createWriteStream(dest);
+    let downloaded = 0;
+
+    const progress = new Transform({
+      transform(chunk: Uint8Array, _encoding, callback) {
+        downloaded += chunk.byteLength;
+        if (contentLength > 0) {
+          const pct = ((downloaded / contentLength) * 100).toFixed(0);
+          process.stdout.write(
+            `\r  Progress: ${formatBytes(downloaded)} / ${formatBytes(contentLength)} (${pct}%)`,
+          );
+        } else {
+          process.stdout.write(`\r  Downloaded: ${formatBytes(downloaded)}`);
+        }
+        callback(null, chunk);
+      },
+      flush(callback) {
+        process.stdout.write('\n');
+        callback();
+      },
+    });
+
+    const readable = Readable.from(res.body);
+    await pipeline(readable, progress, out);
+
+    const durationMs = Math.round(performance.now() - startTime);
+    return { bytes: downloaded, durationMs, contentType };
+  }, url);
 }
 
 // ---------------------------------------------------------------------------
